astro-project/requirements.txt
pandas==2.1.3
apache-airflow-providers-snowflake==5.3.0
astronomer-cosmos==1.6.0
